Moved SampleData file to the DataLayer
