+++ app/src/main/java/dev/baseio/discordjetpackcompose/ui/routes/dashboard/main/dasboard/DashboardUtil.kt
@@ -2,7 +2,7 @@ package dev.baseio.discordjetpackcompose.ui.routes.dashboard.main.dasboard
 
 import dev.baseio.discordjetpackcompose.entities.ChatUserEntity
 import dev.baseio.discordjetpackcompose.entities.server.ServerEntity
-import dev.baseio.discordjetpackcompose.ui.utils.getSampleServer
+import dev.baseio.discordjetpackcompose.utils.getSampleServer
 
 fun getFakeChatUserList(): MutableList<ChatUserEntity> {
   return mutableListOf<ChatUserEntity>().apply {
