+++ app/src/main/java/dev/baseio/discordjetpackcompose/ui/routes/dashboard/serverchannels/ServerChannelList.kt
@@ -10,24 +10,18 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.hilt.navigation.compose.hiltViewModel
 import dev.baseio.discordjetpackcompose.entities.ChatUserEntity
-import dev.baseio.discordjetpackcompose.entities.server.ServerEntity
 import dev.baseio.discordjetpackcompose.repositories.ServerRepoImpl
 import dev.baseio.discordjetpackcompose.ui.routes.dashboard.channels.ChannelList
 import dev.baseio.discordjetpackcompose.ui.routes.dashboard.components.ServerIconSelector
 import dev.baseio.discordjetpackcompose.ui.routes.dashboard.directmessages.DirectMessageList
 import dev.baseio.discordjetpackcompose.ui.theme.DiscordColorProvider
-import dev.baseio.discordjetpackcompose.ui.utils.getSampleServerUIState
 import dev.baseio.discordjetpackcompose.usecases.GetServerUseCase
-import dev.baseio.discordjetpackcompose.utils.UIState
-import dev.baseio.discordjetpackcompose.viewmodels.LandingScreenViewModel
+import dev.baseio.discordjetpackcompose.viewmodels.HomeScreenViewModel
 
 @OptIn(ExperimentalAnimationApi::class)
 @Composable
@@ -35,14 +29,11 @@ fun ServerChannelList(
     modifier: Modifier = Modifier,
     serverId: String,
     chatUserList: List<ChatUserEntity>,
+    homeScreenVM: HomeScreenViewModel = hiltViewModel(),
     openServerInfoBottomSheet: () -> Unit,
     onItemSelection: () -> Unit,
 ) {
-    var serverState: UIState<ServerEntity> by remember { mutableStateOf(UIState.Empty) }
-
-    LaunchedEffect(serverId) {
-        serverState = getSampleServerUIState(serverId = serverId) // TODO: Replace with actual implementation
-    }
+    val serverState = homeScreenVM.currentServerEntity
 
     Surface(
         modifier = modifier.fillMaxHeight(),
@@ -76,7 +67,7 @@ fun ServerChannelList(
 @Preview(showSystemUi = true)
 @Composable
 private fun ServerChannelListPreview() {
-    val vm = LandingScreenViewModel(GetServerUseCase(ServerRepoImpl()))
+    val vm = HomeScreenViewModel(GetServerUseCase(ServerRepoImpl()))
     LaunchedEffect(Unit) {
         vm.getServer("testId")
     }
