+++ data/src/main/java/dev/baseio/discordjetpackcompose/utils/ExtensionUtils.kt
@@ -1,8 +1,12 @@
 package dev.baseio.discordjetpackcompose.utils
 
 import android.content.Context
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import timber.log.Timber
 
 inline fun <reified T> Context.readAssetFile(filePath: String): T? {
@@ -14,4 +18,12 @@ inline fun <reified T> Context.readAssetFile(filePath: String): T? {
         Timber.e(ex, "Some error occurred. Reason:")
         null
     }
+}
+
+fun ViewModel.ioScope(block: suspend () -> Unit) {
+    viewModelScope.launch(context = Dispatchers.IO) { block() }
+}
+
+fun ViewModel.mainScope(block: suspend () -> Unit) {
+    viewModelScope.launch(context = Dispatchers.Main) { block() }
 }
