+++ domain/src/main/java/dev/baseio/discordjetpackcompose/utils/NetworkUtils.kt
@@ -1,0 +1,9 @@
+package dev.baseio.discordjetpackcompose.utils
+
+suspend inline fun <T> safeApiCall(crossinline block: suspend () -> T): T? {
+    return try {
+        block()
+    } catch (e: Exception) {
+        null
+    }
+}
